Fixed bugs with unsigned long transform cartesian to delta
Added some helper math functions
Added some defines
Added capability to define a zone above 0 the carriage should not enter,
as some (Rostock Max) printers carriages will bottom out.
Changed IGNORE_COORDINATE value because we have some printers being
supported near 100k steps, by adding a digit
Added some preliminary stuff to indicate error returns.

+++ src/ArduinoAVR/Repetier/motion.h
@@ -654,7 +654,7 @@ public:
 #if NONLINEAR_SYSTEM
     static void queueDeltaMove(uint8_t check_endstops,uint8_t pathOptimize, uint8_t softEndstop);
     static inline void queueEMove(long e_diff,uint8_t check_endstops,uint8_t pathOptimize);
-    inline uint16_t calculateDeltaSubSegments(uint8_t softEndstop);
+    inline int32_t calculateDeltaSubSegments(uint8_t softEndstop);
     static inline void calculateDirectionAndDelta(long difference[], flag8_t *dir, long delta[]);
     static inline uint8_t calculateDistance(float axis_diff[], uint8_t dir, float *distance);
 #if SOFTWARE_LEVELING
